2차 리뷰 수정
