+++ Hoon/AndroidStudy_Hoon/app/src/main/java/com/study/androidstudy_hoon/SearchActivity.kt
@@ -8,13 +8,9 @@ import android.widget.SearchView
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.study.androidstudy_hoon.data.repository.GithubSearchRepoImpl
-import com.study.androidstudy_hoon.data.service.GithubServiceBuilder
 import com.study.androidstudy_hoon.databinding.ActivitySearchBinding
-import com.study.androidstudy_hoon.domain.usecase.SearchRepoUseCase
 import com.study.androidstudy_hoon.presenter.search.SearchAdapter
 import com.study.androidstudy_hoon.presenter.search.SearchViewModel
-import com.study.androidstudy_hoon.presenter.search.SearchViewModelFactory
 
 class SearchActivity : AppCompatActivity() {
 
@@ -26,7 +22,6 @@ class SearchActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActivitySearchBinding.inflate(layoutInflater)
 
-
         viewModel = ViewModelProvider(this, Injection.provideSearchViewModelFactory())
                 .get(SearchViewModel::class.java)
 
@@ -36,9 +31,8 @@ class SearchActivity : AppCompatActivity() {
     }
 
     private fun initObserve() {
-        viewModel.searchRepoList.observe(this, {
-            viewModel.showProgress(false)
-            searchAdapter.searchRepoList = it
+        viewModel.searchRepoList.observe(this, { repositoryList ->
+            searchAdapter.submitList(repositoryList)
         })
 
         viewModel.isLoading.observe(this, { isLoading ->
@@ -64,7 +58,6 @@ class SearchActivity : AppCompatActivity() {
         val searchView = menu?.findItem(R.id.action_search)?.actionView as? SearchView
         searchView?.setOnQueryTextListener(object: SearchView.OnQueryTextListener {
             override fun onQueryTextSubmit(query: String?): Boolean {
-                viewModel.showProgress(true)
                 viewModel.fetchRepoSearch(query?.trim())
                 return false
             }
